Remove useMenuCheckboxItem hook
